+++ src/qtm4j/resolver/resolvers/test-case-uid-resolver.ts
@@ -45,10 +45,11 @@ export class TestCaseUidResolver extends Resolver {
   ): Promise<Record<string, ResolvedTestCase>> {
     if (keys.length === 0) return {};
 
-    const response = await this.apiClient.get(
-      ENDPOINTS.RESOLVE_TEST_CASE_IDS(projectId),
-      { keys: keys.join(",") },
-    );
+    const response = await this.apiClient
+      .skipAnalytics()
+      .get(ENDPOINTS.RESOLVE_TEST_CASE_IDS(projectId), {
+        keys: keys.join(","),
+      });
 
     return response as Record<string, ResolvedTestCase>;
   }