+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [Swagger] Adjusted some of the Swagger Functional Testing tools responses to return `url` of test or test Suite execution.
+- [QTM4J] Added test execution tools to start executions, update test case and step executions.
+- [QTM4J] Added API request tracking.
+
+### Fixed
+
+- [QTM4J] Fixed `search_test_cases` response validation errors when test case fields are missing.
 
 ## [0.31.0] - 2026-07-16